GameSelectorPanel and BasicGameMenuBar now use Mac OS X native-style open and save dialogs.
Not sure why, but under Mac OS X swing file dialogs don't look right. But, AWT Dialogs do. So, we check for Mac and then display the appropriate dialog.

I didn't change the save dialog in the TripleA menu bar code, mostly because the TripleA menu bar could inherit from BasicGameMenuBar, and then the code wouldn't be duplicated. If this isn't the right approach, please feel free to copy the code from BasicGameMenuBar to the appropriate place in the TripleA menu bar code.

git-svn-id: https://svn.code.sf.net/p/triplea/code/trunk/triplea@1787 a4c711aa-641f-0410-879d-9b464e52b682