fix_mac_icon.sh - Fix DOS end of line character used in script.
The fix_mac_icon.sh script moves an "Icon" file to have that file name end with a dos line ending. In the script the line ending character was incorrectly entered as a "^" followed by "M" rather than the single character generated by typing (on a command shell) "ctrl+v" then "ctrl+m".

Secondly, the script commentary was fixed to not mention carriage return. Dos line ending "^M" is not the same as a carriage return "\r" and should not be confused with each other.