Enable Travis/Gradle for artifact building and automatic push (forked repo first step)
Update gradle to build the all platforms zip and to place release artifacts into an "artifacts" folder. Add travis configuration to build windows installer and automatically push artifacts into github releases (for the moment just a forked repo has the encrypted api-key configuration, next step will be to work on getting the config up for the main repo).

Also scripts were moved into a folder to organize them and make it easier to bundle them.
Last, since Ant build is not being planned for the next release, it is not being updated.