[Image Loading Fix] Fix so images are loaded from Maps first
When images are present with both the game engine, and also a map, then prefer to use the one bundled with the map. Before we would use URLClassLoader and pass it base search paths in a specific folder. It URLClassloader is now no longer deterministic when there are multiple matches, which one it will find is no longer deterministic (java8 upgrade? - who knows. I set up a URLClassLoader that only had the map zip path, and it still returned the wrong one, I was left puzzled why URLClassloader also searched the current path rather than the single one I specified).

To get around this problem we instead ask URLClassloader to return all matching paths and then return any that are in map zip folder first.