Improve Transports To Load Default (#647)
* Improve mapTransportsToLoad to try and fill as many units into as few
transports as possible
* Move transport methods from MoveValidator to TransportUtils
* Clean up method comments
* Remove unused method
* Extract sorting code into helper methods
* Split mapTransportToLoad into 2 methods. One that focuses on finding the
min number of transports needed to load from the list. One that focuses
on distributing units evenly across the list of transports. The reason
this is needed is we want to optimize the default to the min but after
the user selects the number to load we want to put units in all of them.
* Update MovePanel to call the new MinTransports loading method to improve
the default number displayed to the user.
* Organize methods