Inject special map prefix for github 'master' zip files
Use a small ResourceLocationTracker location to contain the logic of if we need a special prefix or not when loading resources. The game previously assumed all map resources files to effectively be 'at the root'. A map downloaded from triplea-maps has a predictable format, it is named: {repoName}-master.zip, with folders "{repoName}/map".

The /map is an addition to better organize map files. We will be required to change the game code anyways to do the map rpefix, and allowing the '/map' means the current repo's do not need to be updated.

So, the Tracker finally will check the name of the zip file we are loading, if it ends with '-master.zip', then we'll assume a special path prefix, otheriwse the path prefix is empty, and things work as they did before.