Move the map download listing back to the game engine
Now that it is simpler to download and play maps (they do not have to be releases, they can be the master branch zip, or a clone), we have available static URLs for the maps. We can use those to reduce the rate of change in the file, which was a concern and reason for moving it. ON the downside of that move is that we had game configuraton scattered, we have a number of files that are already read by the live game in teh triplea-game/triplea repo, so the split of the map config lead to another element that was a bit in left field. In addition, it si possible to override to a local map download config, as a file with the game engine it's easier to use a local version of the file when it is already present on the file system (this makes testing map updates easier).