Replace route translation, fix horizontal line drawing problem
Fixes horizontal line problem reported in: https://github.com/triplea-game/triplea/issues/1056

The problem related to map "seams", where we would go from a high map coordinate, to a low one. For example, a coordinate of 900 to a coordinate of 50 with a map width of 1000. The existing code would translate and get at least the mouse cursor point to be effectively 1050. The next point though could be something like 70. It seems in that case two far offscreen coordinates were used to draw the route "the wrong way" around the globe. This effectively comes out to be a flat line.

For the fix, it was observed that routes wtih negative coordinates worked well. In this case we use the map width to detect if we have crossed a map seam. If two points are further apart than the map width, then we must have crossed it. For any points that are greater than a mapWidth apart, we subtract the full map width to get a negative number. This way for example a point that was once at postion 900 on a 1000 width map is now at -100. So for example when previously we may have had x positions 900 and 50, that would translate to -100 and 50.

Note, this does not fix the vertical direction, it might be broken there with weird vertical lines.