Remover 'faster' scroll zone on map edge (includes bug fix)
The faster scroll zone divided the map edge scroll zones into two. The first one, further away from the edge, would scroll at a fixed rate, and then when closer in the faster scroll the scrolling would  speed up. The feature never caught on very well, and was a substitute for the lack of controls over the scroll zone size and scroll speed. With those as features, simplifying this not popular feature seems prudent.