Use a cache when querying download length, do not start download progress bars until download starts.
Should fix download progress bars that display "[ ]". The problem there is that we would request
the length of each download file all at the same time while setting the upper length of each progress
bar. To fix, we move the set of the upper end length to the time when download starts. We download 3 maps at the same time, which means we'll fire off a batch of 3 requests.

To further avoid request throttling, we can add a cache to avoid unnecessary requests for the content length of URLs which we have already requested once before.