Apply Swing Look and Feel immediately to mitigate poison pill look and feels
When look and feel is selected, it is saved to preferences and then loaded next on startup. It was reported in: https://github.com/triplea-game/triplea/issues/1208 a case where startup froze while applying the look and feel. This then required the user to find the system preference file on their system and to update it by hand to avoid the error on every startup (poison pill).

To mitigate this, we can apply the look and feel setting immediately, before we set the system property. This way if we never get past the look and feel, we will leave the system property unaltered. At least in this case if TripleA is force killed, the system property will remain a valid default.

On one hand it is very nice the look and feel is applied right away. On the other rendering is very weird. So, a simple messaging update is included to let a user know that the setting has taken partial effect, and recommends a restart.

(while in that code the setting option delayed map parsing was tweaked  as well to explicitly indicate it can hurt game performance)