Fix resource not found errors in release build
When running TripleA from a release build (e.g. the result of the
allPlatform task), numerous resource not found errors occur.  The root
cause is due to the resources being located in the
src/main/resources/data folder instead of src/main/resources.  The
presence of the data folder prevents calls to Class#getResource() that
specify a relative path from succeeding because the inferred absolute
path does not exist on the classpath.  The fix is to move the contents
of src/main/resources/data to src/main/resources.

These errors did not occur when running TripleA from Eclipse because the
various launch configurations explicitly included the
src/main/resources/data folder at the root of the classpath.