Prevent UI deadlock after version update (#1493) (#1558)
* Show message dialog immediately if already running on the EDT

This change prevents a deadlock that will occur if this method is
invoked from the EDT.  In that case, the method will block waiting for a
latch to be signaled.  However, the code that signaled the latch is
forked in a SwingUtilities.invokeLater() call.  Thus, the method is
awaiting a signal that will never occur because it is blocking the EDT
from processing new events, which includes the forked latch count down.

* Extract method to handle running code on EDT and waiting from any thread

The code in the various showXXXDialog() methods was fairly common except
for the code that actually called into JOptionPane.  This change removes
the duplication and isolates the core logic for invoking some arbitrary
code on the EDT and awaiting it (with CountDownLatchHandler support)
from any thread, possibly including the EDT.

* Replace smelly IGNORE return value with Optional<>

* Remove Travis Xvfb requirement

Even though our unit tests spin up an EDT in the headless environment on
Travis, they do not perform any operations that may cause a
HeadlessException to be thrown (e.g. creating a Swing component).
Therefore, running Xvfb on Travis is not necessary.

* Add test to cover invokeAndWait(..., IntSupplier) overload