Fix game.dtd not found errors in release build
This change partially reverts 3730dd8.  It makes game.dtd available to
GameParser so attempts to load a game XML file from a release build do
not fail.

These errors did not occur when run under Eclipse because the game.dtd
file was located in the project root, which is the CWD for the Eclipse
launcher used to run the application.  Thus, the parser was able to
locate the DTD and load it.