Prompt user to download tutorial map at startup if applicable
The user will be prompted to download the tutorial map if all of the
following conditions are true:

- The system preference TRIPLEA_PROMPT_TO_DOWNLOAD_TUTORIAL_MAP is set
  to "true" or is not present.
- No maps are present in the user maps folder.

After checking the system preference
TRIPLEA_PROMPT_TO_DOWNLOAD_TUTORIAL_MAP, its value will always be set
to "false".  This prevents the user from being prompted to download the
tutorial map in the following scenario:

1. User has a version of TripleA installed prior to this change.
2. There is at least one map in the user maps folder.
3. User installs new version of TripleA that contains this change.
4. User is not prompted to download tutorial map because the user maps
   folder is not empty.
5. User subsequently empties the user maps folder.

If we didn't set the preference value to "false" after step (4), the
next time they start TripleA, they would be prompted to download the
tutorial map.  This seems awkward because the user is most likely not a
brand new player.