Checkstyle: Document Unicode escape usage
The Google style guide recommends that printable Unicode characters be
entered directly into source files instead of using Unicode escapes.  In
this case, the code points are non-printable combining characters.
Thus, usage of a Unicode escape is acceptable.  The Checkstyle warning
is suppressed by adding a tail comment to document the escape usage.