Upgrade to Mockito 2.7.22
The following types were deprecated since Mockito version 2.0.82-beta
and had to be replaced with equivalent types:

* o.m.Matchers => o.m.ArgumentMatchers
* o.m.runners.MockitoJUnitRunner => o.m.junit.MockitoJUnitRunner

Since Mockito 2.3, "strict stubs" [1] are enabled by default.  This
feature seems beneficial, so instead of disabling it, I went ahead and
removed the unnecessary stubbing that was causing three tests to fail.

The test RouteTest#testCorrectParameterHandling() began failing after
the upgrade due to an NPE:

java.lang.NullPointerException
	at games.strategy.triplea.ui.logic.Point.<init>(Point.java:22)
	at games.strategy.triplea.ui.MapRouteDrawer.getRoutePoints(MapRouteDrawer.java:214)
	at games.strategy.triplea.ui.MapRouteDrawer.drawRoute(MapRouteDrawer.java:71)
	at games.strategy.triplea.ui.RouteTest.testCorrectParameterHandling(RouteTest.java:91)

The root cause was due to a combination of issues:

1) Since Mockito 2.1, the any(Class) matcher no longer matches null.
2) The stubbed Route had its m_start field set to null, which resulted
   in the Route#getAllTerritories() method returning a list with a null
   element.

Therefore, the null Territory was not matching the stub for
MapData#getCenter(Territory), and the default Mockito stub returned
null, which triggered the NPE above.

The fact that the Route has m_start set to null is a real problem and
was being covered up by the old Mockito behavior.  Therefore, the proper
fix was not to simply add another stub for MapData#getCenter() in the
case of a null Territory value.

The m_start field was not set because the test creates the Route using
the default constructor and then adds territories via the Route#add()
method.  Route#add() does not change the m_start field, and assumes the
caller will properly initialize it via Route#setStart().  This was never
done in the test code.  The fix was to use the overloaded Route
constructor that accepts a list of territories, which properly
initializes both m_start and m_steps.

[1] http://static.javadoc.io/org.mockito/mockito-core/2.7.22/org/mockito/junit/MockitoRule.html#strictness(org.mockito.quality.Strictness)