Initiate download of out-of-date maps upon user request
Previously, when the user answered Yes to the prompt to update their
out-of-date maps, we would simply open the Download Maps window, and the
user would be required to manually update the maps.  This change
initiates the download of the out-of-date maps automatically in addition
to opening the Download Maps window.