Prevent requesting download length multiple times
A ListSelectionListener may receive a chain of events for a single
selection change.  The last event in the chain is distinguished by
the valueIsAdjusting property being set to false.  Therefore, we should
ignore all but the last event in the chain to avoid performing the same
action multiple times for the same logical change.