Do not update progress bar until at least one byte has been received
FileSizeWatcher monitors the length of the temporary local file.  That
file will be created almost instantaneously compared to when the first
byte is received from the network.  Thus, the progress listener will
receive several updated events with the current length set to 0.

This causes the indeterminate status bar to disappear almost
immediately, and it is replaced with the standard 0% display.  That kind
of defeats the purpose of having the indeterminate progress bar.
Therefore, with this change, the behavior is to display the
indeterminate progress bar until both

1. the total download length is received, and
2. at least one byte of the download has been received.