Fix possibility of multiple progress bars per download
The problem was due to moving the call to
DownloadListener#downloadStarted() from
DownloadFile#startAsyncDownload() to DownloadCoordinator#accept().

When called from DownloadFile#startAsyncDownload(), downloadStarted()
will only be invoked once per download.  However, the code in
DownloadCoordinator#accept() was calling downloadStarted() for each
listener every time a new download was started.  This had the side
effect of creating additional progress bars for in-progress downloads
whenever a new download was started.

I simply restored the previous code where downloadStarted() is called
from DownloadFile#startAsyncDownload().  If there was a reason this code
was moved to DownloadCoordinator#accept(), it should be re-evaluated.