Redesign how download events are delivered and handled
The previous code always assumed a download started event was guaranteed
to be delivered once.  I was able to exacerbate a race condition and
observed DownloadCoordinator#accept() being called twice for the same
download.  This caused multiple downloads to be started for the same URI
and multiple progress bars to appear.  The earlier progress bars would
stop reporting progress altogether, while the latest progress bar's
value would fluctuate up and down between multiple values as the
download progressed.  The root cause of the progression was that the
on-disk file lengths for multiple files were being sent to the same
progress bar.  As those files had different lengths simultaneously, it
would appear that the progress bar value both increased and decreased
over time.

The primary change made here is to ensure DownloadCoordinator will not
accept a download request for an URL that is either pending or active.

A secondary change is to redesign the code to handle the case where a
download started is not delivered.  That is, if the first event a
listener observes is an updated or completed event, it should just work.
The non-delivery of a download started event now occurs when opening the
Download Maps window with one or more downloads in progress.  For those
maps, a download updated event will be the first observed event.

The new thread created in response to the Install button being clicked
was removed as it didn't appear to be necessary.  In addition, the JList
should probably always be accessed from the EDT (even though in this
case, calling getSelectedValuesList() is apparently safe to call from a
non-EDT thread).

Finally, the DownloadListener#downloadStarted() method is now called
from the thread created by DownloadFile#createDownloadThread().  This is
to make it symmetric with the corresponding call to
DownloadListener#downloadStopped() in the same method.  Because the
downloadStarted() and downloadStopped() methods are called from one
thread, while the downloadUpdated() method is called from another thread
(see FileSizeWatcher), there is the possibility that downloadUpdated()
will be called before downloadStarted().  However, the code has been
re-designed to handle the possibility of methods called in this order.