Do not throw exception upon failure to protect password
Throwing an exception during serialization will corrupt the entire local
bean cache, forcing it to be cleared.  This is an undesirable side
effect because it could possibly delete other information unrelated to
the bean containing the password.  Instead, we log the exception and
clear the password.  The end result is the user will have to re-enter
the password associated with this one bean.