Migrate all but DBUser.java from 'userDB' package to 'userdb' (#2008)
* Migrate all but DBUser.java from 'userDB' package to 'userdb'

There are compatibility problems if we move 'DBUser.java', but we are free to move the other class files.
'userDB' is a checkstyle violation for package naming, it makes working with these classes difficult. Adding a test for example will add a checkstyle violation since there is a new (test) class in a badly named package. In this update we move the classes we can to reduce this problem.

* Fix import orderings

* Move all in package 'userDB' to 'db' except DbUser.java. Avoids checkstyle issues.