Suppress varargs linter warnings
These methods do not perform unsafe operations on their varargs
parameters.  They simply pass them on to Arrays#asList(), which is
varargs safe.

While Eclipse is satisified with just the presence of the @SafeVarargs
annotation, javac (both Oracle and OpenJDK) report additional varargs
warnings on these methods.  Following the example of Arrays#asList(), it
is sufficient and correct to also add @SuppressWarnings("varargs") to
these methods.

Unfortunately, "varargs" is currently an unknown type of warning to
Eclipse, and the @SuppressWarnings("varargs") annotations will be marked
as such by the IDE.  There is an open Eclipse bug [1] [2] to add support
for suppressing "varargs" warnings.  In the meantime, Eclipse users will
have to be content with the reported unsupported warnings.

[1] https://bugs.eclipse.org/bugs/show_bug.cgi?id=349669#c21
[2] https://bugs.eclipse.org/bugs/show_bug.cgi?id=344783