Revert potentially unsafe changes from #1988
Non-UI classes may be accessed concurrently.  Thus, it may not be safe
to change instances of CopyOnWriteArrayList to ArrayList.  Further
review is required to determine if any of the class fields modified in
this commit are, in fact, not accessed concurrently, or are protected by
a lock if accessed concurrently.  They then can be changed back to
instances of ArrayList.