Fix Travis build break
We do not specify an installation directory for install4j, therefore the
"default" installation directory is used.  On the previously-used
Precise image, this apparently was "~/install4j6".  However, after
upgrading to the Trusty image, it is now "/opt/install4j6".  Our scripts
are hard-coded to expect the former location.

The fix employed in this change is to explicitly specify the install4j
installation folder via the "-dir" command line switch.