Use helper to register window closed action
SwingComponents#addWindowCloseListener() was a little misleading because
it actually listens for the "closing" event instead of the "closed"
event.  DownloadMapsWindow specifically needs to listen for the "closed"
event, so I added a new method for this case, as well as renaming the
existing method so it is clear it is listening for the "closing" event.