Remove incremental build support for prepareGameEngineProperties task
The prepareGameEngineProperties task not only depends on the
game_engine.properties file but also the value of the engineVersion
project property.  The task only tracks the state of the former and not
the latter.  Thus, if a dev rebuilds with the same
game_engine.properties file but a different engineVersion project
property, the output file will not be regenerated.

This is such a small and fast operation that incremental build support
was simply removed and the task will be run on each build.