Restore RMI compatibility
PR 1766 introduced a feature that required the addition of a new method
on ITripleAPlayer in order to prompt the user to select the territory to
use for production when multiple factories are adjacent to a sea zone.
Unfortunately, adding a new method to a remote interface changes the
ordering of the methods, as calculated by RMI, which causes a mismatch
between method calls between newer and older clients.

I could not find a way to preserve the feature from PR 1766 while
maintaining RMI compatibility.  Therefore, this commit reverts most of
PR 1766.  The user can no longer select the territory to use for
production when there are multiple possibilities.  Instead, the engine
simply uses the first available territory.  Thus, the problem described
in issue 350 has been re-introduced.