Re-enable attachment validation tests (#2136)
* Fix broken ValidateAttachmentsTest field check

The PropertyUtil#getFieldIncludingFromSuperClasses() method only looks
for the field name as given.  It does not also try looking up a field
prefixed with "m_" as the PropertyUtil#getPropertyField() method does.

To fix this, a new getPropertyField() overload was added that accepts a
Class in place of the Object parameter in the existing overload.  The
two calls in the codebase to getFieldIncludingFromSuperClasses() were
replaced with this new overload.  The ValidateAttachmentsTest code now
runs the same logic as the production code.

* Add missing newline in error message

* Fix invalid CanalAttachment#getExcludedUnits() method

An attachment property getter must not accept any parameters.

* Add missing attachment getter methods

* Add missing attachment resetter methods

* Add support for virtual game properties

A virtual game property either has no backing field or uses one or more
backing fields associated with other non-virtual properties.  A virtual
property is typically used for so-called "convenience" properties that
are used to initialize multiple properties at once.  Properties of this
type are not required to have get, reset, or clear methods.

* Include ResourceCollection in list of valid field types for adders

* Add missing attachment getter methods

In these cases, the getter methods were present but didn't conform to
the attachment requirements because their return type differed from the
associated field type.  In fact, these getter methods were actually
convenience methods whose return value was derived from the property's
backing field.

To fix these violations, the existing methods were renamed to something
more meaningful (usually by simply removing the "get" prefix from the
method so that they read more like a predicate).  Then new getter
methods with the appropriate name were added that simply return the
value of the backing field.

* Fix broken testAllAttachments test

* Add missing annotations and resetter method for TestAttachment

* Fail tests when attachments are invalid

* Skip testAllAttachments in a headless environment

This test attempts to load pretty much every TripleA class, including UI
classes, which will fail to load in a headless environment.

* Update list of known attachment classes