Add serializable proxy for ProductionFrontier
There was a lot of refactoring in the Matchers class to ensure all
significant fields are considered when comparing two instances for
equality.  A problem was identified whereby some superclasses (e.g.
DefaultNamed) implemented equals(), but the implementation did not
consider all significant fields.  This caused false positives during
testing.

For legacy reasons, the equals() implementations cannot be changed.
The workaround is to register custom "equality comparators" for
each class and to provide a matcher that performs a deep equality
comparison by delegating to the custom equality comparators.  This is a
little verbose at the moment, and some refactoring is envisioned in the
near future.