Add '@InternalDoNotExport' to satisfy ValidateAttachmentsTest checks.
Without, 'ValidateAttachmentsTest#testAllAttachments' fails with:

```
Could not find attachment 'this is surely not a valid identifier'. This is can be a map configuration problem, and would need to be fixed in the map XML. Or, the map XML is using a feature from a newer game engine version, and you will need to install the latest TripleA for it to be enabled. Meanwhile, the functionality provided by this attachment will not available.
Could not find delegate 'this is surely not a valid identifier'. This is can be a map configuration problem, and would need to be fixed in the map XML. Or, the map XML is using a feature from a newer game engine version, and you will need to install the latest TripleA for it to be enabled. Meanwhile, the functionality provided by this delegate will not available.

java.lang.AssertionError: One or more attachments are invalid:
WARNING: Class games.strategy.engine.data.FakeAttachment setter setName: begins with 'set' so must have either InternalDoNotExport or GameProperty annotation
```