Exit application when Quit is selected from main menu
We now explicitly terminate the application when the user selects Quit
from the main menu.  Previously, we only disposed the main frame, which
would allow the application to continue running if there were any other
frame windows open (which was the case when the error console was open).

This behavior is identical to the in-game Exit Game command, which
explicitly calls System#exit() to terminate the application.