Extract helpers for reading from and writing to in-memory streams
Introducing these helpers is intended to reduce the confusion over
whether or not ByteArrayInputStream and ByteArrayOutputStream need to be
closed.  Prior to this change, code was inconsistently split between
not closing and closing (usually through try-with-resources).

They also help to ensure filter output streams around an in-memory
stream *ARE* closed before subsequently reading from the resulting
buffer.  If the filter output streams are not closed before reading from
the buffer, there is a possibility that reading from the associated memory
input stream will end prematurely.