Fix JavaMail warnings in PBEM games
Because we only require SMTP support, we only link with the JavaMail
mailapi and smtp JARs.  Neither of these JARs contains the following
files:

* /META-INF/javamail.default.providers
* /META-INF/javamail.default.address.map

This causes the JavaMail Session implementation to log two warnings,
which causes the Error Console to be displayed unexpectedly whenever the
user emails a save game.

The fix is to use the JavaMail javax.mail JAR instead of the mailapi and
smtp JARs.  The javax.mail JAR contains the required files, and thus no
warnings are displayed when emailing a save game.  Note that the
javax.mail JAR also contains the POP and IMAP transports, so it adds
226 KiB to the size of the distribution.

An alternative solution would be to continue using the mailapi and smtp
JARs but provide our own copy of the javamail.default.* files.  This
solution was not chosen because it adds additional code to be
maintained and a slight increase to the complexity of the build process.
A less than 226 KiB increase in installer size was seen as a reasonable
trade-off to the additional complexity.