Fix IllegalStateException when parsing maps
Certain code paths may call GameChooserModel#newInstance() from a thread
other than the EDT.  In such a case, an IllegalStateException is thrown
because this method assumes it will always be called from the EDT.

The fix is to determine if the current thread is the EDT.  If so, use
BackgroundTaskRunner to parse all maps.  Otherwise, parse all maps on
the current thread.