Fix chat message history navigation keys
The values in a Swing component's InputMap must be either a key that has
a corresponding entry in the component's ActionMap or it must be an
Action itself.  ba8a9e6 switched the Action values to ActionListeners
for the message history navigation keys.  This had the effect of not
running any action when these keys were pressed because ActionListener
values in the InputMap are not directly executed and there was no
corresponding entry in the ActionMap.