Download Substance dependencies from GitHub during build
The dependencies must be stored in a folder other than "build/",
otherwise running the Gradle "clean" task in conjunction with any other
task will fail because the downloaded dependencies will be removed after
they have been downloaded.  The chosen solution is to store the
downloaded dependencies in a folder sibling to "build/"
(".remote-libs/"), and to provide a separate task ("cleanRemoteLibs") to
remove this folder independently of the "clean" task.