Remove ClientLogger usage from tests
In some cases, the exceptions that were being swallowed and logged
should be allowed to raise a test error.

For the cases that quietly shut down an IMessenger, the logging was
retained, although this is questionable, as the output of an automated
test is typically ignored.  Regardless, the duplicate code used to
quietly shut down an IMessenger was extracted to a new method.