Ensure ErrorMessage is always initialized on EDT (#2925)
* Ensure ErrorMessage is always initialized on the EDT

ErrorMessage initialization must occur as early as possible in the
process.  If ClientLogger#logError() is called before ErrorMessage is
initialized, the initialization may occur off the EDT (e.g. when a
thread's unhandled exception handler is used to process an unhandled
exception).

* Display exception type if message is null

Otherwise the error dialog will display "Error: null", which is not very
helpful.

* Remove outdated comment

The scenario described in this comment no longer appears applicable with
the latest Substance library.