Record all registered and anonymous lobby logins in database (#2977)
* Record lobby login metrics

* Move MockitoExtension annotation to outermost fixture

* Use LocalDate instead of Instant

* Record individual user lobby access

Instead of recording lobby statistics (e.g. number of logins per day),
we now record each individual successful access, both for authenticated
and guest accesses.

* Change "login" to "access"

* Change "authenticated" column to "registered"

We are not modelling whether or not the user was authenticated because
"anonymous authentication" is authentication.  Rather, we are modelling
the type of authentication method used.

* Use "Authenticate" suffix instead of "Access" in AccessLog methods

* Pass enum to AccessLogDao#insert() instead of boolean

* Rename AuthenticationType to UserType

* Revert d75e3ddec

IntelliJ still requires the Mockito annotation on the leaf classes in
order to recognize them.