Restore AbstractRulesAttachment#setTurns(String) method (#3102)
Some maps (e.g. WW2 Europe 1940 Original) trigger a change to the
"turns" property of an attachment during the game.  As the trigger is
coming from the XML, the property value will be a String.  If the code
can't find a property setter that takes a String, it will take whatever
setter it can get.  The only other overload available is one that
accepts a HashMap<Integer, Integer>.  Calling this overload results in
an exception due to an argument type mismatch.

The fix is to restore the setTurns(String) overload, which was renamed
to setRounds() in 0cd3154a5.  The restored overload simply delegates to
setRounds(String), as that method is still required for map XML
properties that use the "rounds" property (which is simply an alias for
the "turns" property).