Improve tooltip manager to hide unit tooltips when the window (#3499)
becomes inactive - e.g. because a dialog was opened.

To do this properly, we need to listen for both window events
and ancestor events - as the parent window may change - and in
particular because this is the case when the component is first
created - when it doesn't have a window yet.