Remove setWidgetActivation callbackmechanism on main screeen and enable buttons by default. (#3569)
MetaStartUpPanel disables a few of the start game button options on startup and is triggered by a callback mechanism to enable those buttons. It turns out the screens that are disabled can handle a not yet initialized game data object. There are a few NPE locations to deal with, but once dealt with we can enable the buttons at startup and remove the setWidgetActiviation callback mechanism.

This update has the benefit that the user will not need to wait for their default map to load before they can click single player, PBEM or host network game; those buttons will be immediately enabled by default.