PR template - Uncomment template!  Add a 'functional changes' section and other cleanups. (#3586)
* Update PULL_REQUEST_TEMPLATE.md

Notably uncomment the template. This means if we fill in the template we will not need to do anything more, like for example double check that the comment blocks have been removed : )

Added a 'funcational changes' section for us to call out if something is purely refactor or not. Making this clean in the PR overview should help review.

Also of note the documentation links for dev were pruned down in text and pushed to bottom and map maker links would be moved to triplea_maps.yaml: https://github.com/triplea-game/triplea/pull/3585

* Tighten text up and add a 'before' and 'after' section for screenshots