Dev quality of life update - LobbyRunner finds config from current location (#3602)
* Dev quality of life update - LobbyRunner finds config from current location

Add a secondary search location when looking for lobby property file. This is intended to help development. Current lobby runner does not work out of the box. There is a compromise here as we could update the build system; here we take the choice to add a code update to check for a second location.

* Use "#orElseGet" to avoid parameter evaluation, fixes up tests