Add bot startup check for map folder. (#3632)
* Add bot startup check for map folder.

If map folder does not exist then bot will not function properly. The bot starting can give us a false sense that things went well. This update adds to the startup checks that the maps folder exists and fails startup if it does not. This should better get our attention when we have a significant misconfiguration.

* Formatting (fix checkstyle)