Fix unused field static analysis warning (#3651)
The MapUnitTooltipManager#deactivationObserver field was previously
unused.  Simply removing the field and creating the
WindowDeactivationObserver instance in the constructor would have fixed
the unused field warning but would have introduced an unreferenced
object warning.

Instead, I reversed the responsibility for managing the
observer/listener.  Instead of the observer adding/removing itself from
the subjects, the outer (owner) now takes on this responsibility.  This
causes the observer field to now be required.