Fix race condition in tests using ClientSetting (#3771)
ClientSetting#saveAndFlush() runs flush() asynchronously.  Test cases
that use ClientSetting invoke ClientSetting#resetPreferences() when the
test is complete.  Depending on timing, this may cause flush() to throw
an exception because preferencesRef will contain a null reference.

The fix is to save the Preferences before spawning the new thread and
passing it to the thread closure.