Disable JDK linter annotation processing warnings (#4042)
This class of warnings only seems useful if we were writing our own
annotation processors in order to detect mistakes that prevent the
processors from being applied to the desired annotations.  Since we have
no annotation processors in the codebase, the warnings we get about
third-party annotations not being processed is simply noise.

It would be ideal if the JDK provided the ability to suppress this
warning for individual annotations that are known to not require an
annotation processor.  Unfortunately, it doesn't.