Replace GameEnginePropertyReader with ProductConfiguration (#4048)
* Replace GameEnginePropertyReader with ProductConfiguration

* Change "qualifier" to "buildId"

* Set Gradle "buildId" project property globally for all Gradle runs

* Remove unnecessary touch

* Remove outdated comment

A release is already in process at the time the push_tag script is run;
pushing a tag does not trigger a new release.

Unlike a normal Travis build, pushing a tag does not trigger a release in
the TripleA build. Rather, the TRIPLEA_RELEASE environment variable and
the repo/branch identity trigger a release.   See the "deploy.on" section
of the build script for more information.