Fix determination of transported units (#4263)
The TransportTracker#transporting(Collection<Unit>, Collection<Unit>)
method was mistakenly removed in #3931 and replaced with
transporting(Collection<Unit>).  This had the unintended side-effect of
incorrectly determining which units are being transported during a
retreat from an aborted amphibious assault.

The fix is to restore the method removed in #3931.  It has been renamed
to transportingWithAllPossibleUnits(Collection<Unit>) to distinguish it
from the original transporting(Collection<Unit>) method because it now
has the same signature due to the removal of the unused first parameter.