Add offline lobby (un)mute (#4468)
* Add support for offline mute/unmute by username

Need to guard against negative delays when scheduling timer tasks.  It's
possible for the clocks between client and server to be slightly
out-of-sync, which may generate a negative scheduling delay, which will
result in an exception on the server.  (This is even possible currently
when specifying a zero duration to the Quick Unmute command.)  Solution
is to round all negative delays to zero so the task is immediately
scheduled.

* Extract helper to create unban timer task

Also inline helpers related to creating unmute timer task, as they are
only used by one caller each.

* Extract helper to validate username entered by user

* Move separator to correct menu

* Extract helper to validate hashed MAC address entered by user

* Annotate nullable instants and document in Javadocs

* Extract helper to calculate timer delay