Return Optionals on LobbyServerProperties & Enforce Port to Be Set (#4522)
* Return Optionals on LobbyServerProperties & Enforce Port to Be Set

- Update LobbyServerProperties API to return `Optional<String>` when
data is not set rather than return an empty string. A 'set' error
message is now the indicator if server is not available rather than
there being a specific method for it.

- 'int' -> 'Integer' is to allow the '@Nonnull' annotation to be applied
and force the property to be set as part of builder construction. Otherwise
we could omit the 'port' from the construction and '#getPort' would then return
'0', a silent error.

* Make httpHostUri non-nullable