Update/Simplify bot launching CLI params (#4545)
* Simplify bot launching

- Simplify: we have some confused code that checks for required and
non-required args, then after that we have some long-time code that
checks the same non-required args are then required. This first check
was completely removed.

- Simplify: non-required args are removed minus bot password (pending
more investigation if it is used). YAGNI on this removal to simplify to
what we use, which is just the required set.

- If the triplea.name does not start with 'Bot', we will now prepend
it first rather than crashing

- remove lobby-hosted-by, we require that value to match 'triplea.name'
exactly, we may as well just use 'triplea.name' only

- comment is now fixed as 'automated_host' and no longer an arg

- server=true is removed and simply hardcoded, same with default
reconnect time

- updated IDEA and eclipse launchers to have updated params

* Fix partial math, "21600" is precomuted but "2*21600" is not precomputed

* Provide default for maps folder, do not check it exists

In `ClientFileSystemHelper.java` we'll create the download map folder if
it does not exist, so we do not need to check that it exists on startup.

Second, map folder override can be defaulted to the standard download
folder map for dev convenience.

* Make lobby reconnect time match prod

In prod we have the reconnect time set to 172800 seconds (or 2 days):
https://github.com/triplea-game/infrastructure/blob/master/roles/application/bot/files/run_bot.sh#L57