Display log level on log level button (#4607)
* Display log level on log level button.

- Primarily we update the console log level button to display the
current log level. This is to provide feedback to the user of what is
selected without having to click on the log level button.

- Text on log level button updated to be short and not have the down
arrow. The down arrow does not always render, hence "Log Level []"
looks a bit confusing, as if there were a 'null' or no log level is set.

- Remove error checking of log level setting. A user cannot set the
value to an arbitrary text, it should be valid.

* Refactor console window to have a view model

Functionally the log level button is now a drop-down box. Code-wise
console window now has a 'view-model' which references a 'ConsoleView'
that is implemented by 'ConsoleWindow'. This removes most of the logic
in 'DefaultConsole' which is now 'ConsoleWindow', and keeps swing logic
out of the model code in 'ConsoleModel'.

* Inject testable clipboard, break circular dependency

1. break console model and view circular dependency:
 - to avoid the view from creating a model, the model is injected into
the view
 - to avoid the model from having a dependency on the view, the view
object is passed to methods that need a view.

2. Inject a strategy object for setting system clipboard into console
model.

3. Move window visible responsibility from console view to model.

4. Split log handling initialization code to respective components