Update: Error Message window LogRecord print handling (#4680)
* Update: Error Message window LogRecord print handling

Error message window was coded to expect LogRecord.getMessage() to
always contain a value. If we have an uncaught exception with no
message, eg: `new NullPointerException()`, then the error message window
is blank.

This update conditionally prints the log message if it is there, the
exception message if it is there, and will not double print the message
if the log message and exception message match (which happens in the
case of an uncaught exception with a message, eg: `new
NullPointerException("message")`.

In addition to message printing we'll always have a third line printed
'location' that will tell us where the occurred. This serves a few
purposes:
- nice for dev, we get a pretty complete picture of the error just from the messages
and where the error occurred
- for user, it is a way to give feedback in case of uncaught errors and
limited information whether we are serving the exact same error or
different.

* Simplify format, remove "location line" and details/error prefix

* Test cleanup: inline test object and remove commentaries