Wire up client to send errors to http server / Change 'shows details' button to 'report to TripleA' (#4687)
* Add 'toolTip' property to JTextAreaBuilder

Enables a swing property in JTextAreaBuilder that will be used
in an upcoming change.

* Method name update, borderEmpty -> border

The 'border' method now becomes an overload, bit more obvious that it is
a simplified version of 'border(BorderBuilder)'

* Wire up client to send errors to http server

User facing change where on the error message pop-up window, instead of
having a 'Show Details' button (that would open the console), we now
have a 'Report to TripleA' button that opens up a form window. The form
window has one text field where a user can enter details (optional), and
click 'upload' to then send to http server.

Notable code/design changes:
- LogRecord maybe having an exception is a bit tricky, there is a lot of
logic in formatting the title+body to have data extracted from log
record.
- Error report window is essentially re-written, focusing on case where
we have a LogRecord object and user input is optional.
- Error report window model has new design/contract where it interacts
with the UI only through interface. The UI has a 'bind' method to let it
attach model methods to UI components. This avoids a circular dependency
between UI and model, and model and the view interface have no JavaFx or
Swing dependencies, so we can swap out different view interfaces between
the two without changing model.

* Simplify: Use Collectors.joining over String.join

* Change: Single space the stack trace in error report instead of double space