Fix chat panel not rendering on game staging screen. (#4782)
* Fix chat panel not rendering on game staging screen.

Fixes: https://github.com/triplea-game/triplea/issues/4761

Relatively surgical fix of an impossible `instanceof` check condition that prevented the chat panel from rendering The condition was impossible as the class hierarchies did not line up.

One fix considered was to update the instanceof check to check for presumably the right chat model type 'HeadlessChatModel', but this still would have left in the brittle instanceof check. The fix here rolls back to an explicit 'isHeadless' method that was previously used to do this check.

* Rename param

* Instead of checking 'isHeadless', request the rendering componet.

This will allow us to decouple a model that is also its own view. After this update the model component returns an optional 'Component' that is the view. Currently it returns 'this', which avoids casting, and would allow us to change the object returned in the future without having to any code that currently uses 'ChatModel'