Add DB deployment to ansible and support scripts (#4841)
* Add DB deployment to ansible and support scripts

Updates ansible deployment scripts to be more functional and can now
deploy a functioning lobby server with database.

Key changes:
- Adds database installation to ansible. Previously we had to install Postgres
by hand and configure the database user and create the database by hand. Now
ansible will do this.
- Adds flyway deployment to ansible.
- Database user is changed from 'postgres' to a new database user 'triplea-lobby'.
This is done to follow best-practices and not use the postgres DB super-user account
for application access. This update required changes to the flyway migrations.
When deploying to production we will need to do a fresh migration and a table data
copy to account for this. The intent is for the next production deploment will happen on a fresh
set of servers and fresh database for a less risky cut-over.
- Concept of infrastructure server is introduced, this a single server from where we will
run ansible. The setup on this server is one-time. Copies of scripts deployed to this server
are checked in under 'infrastructure-machine' and should be kept in-sync. A notable
TODO is to update travis to execute 'run_deployment' for prerelease on this server after
successful merges.

Notable Changes:
- Product version is now set via environment variable in ansible
- A secrets file is now deployed to an infrastructure server. The secrets
file is a series of export statements that will set environment variables
from where we can grab password values in ansible scripts.
- Admin user is renamed in ansible scripts from 'triplea' to 'admin'. It was
confusing as who the admin user was, and currently we are using the user 'admin'.
With this change the 'triplea' user on systems would no longer be used. Each app
like lobby and bot have their own system users created already.
- Http server and lobby have been merged, this update reflects that and merges
http_server configuration into lobby.

* Temporary, keep SQL table ownerships with postgres user until Docker is updated.