Enhance swing builders: JButton, JFrame, JTextField, SwingAction (#4864)
Add some additional features to Swing wrapper interfaces. For example
we add some convenience methods to the frame builder so it can be set to
visible from within the builder.

Note: these updates are forward looking and are planned to be used in future updates
(currently used by in-flight work only).