Fix some mixed mutability error prones (#4879)
* Fix some mixed mutability error prones

Fix a number of error prone warnings for mixed mutability on return types:
```
warning: [MixedMutabilityReturnType] This method returns both mutable and immutable collections or maps from different paths. This may be confusing for users of the method.
```

It looks like a number of these are from returning Collections.emptyList() and also returning array list. The former is immutable and the latter mutable.