Fix: Handle save game file not found (#5014)
User is able to select a missing save game from the save game dialog. When this happens we would assert the file exists. It is not guaranteed and we wind up with 'control-flow-by-exception-handling.' The problem with the exception is we offer the choice for a user to report the bug to TripleA when it is not a bug.

This update checks for missing file rather than asserting, and if so notifies the user through an error dialog that the file does not exist