Add: Update unit_scroller to have a unit_sleep button (#5036)
Unit sleep will auto-skip a unit until it has been moved. Useful for example for units that are on islands or AA guns.

In this update the old unit-skip image, a moon, is renamed and re-used to be the unit sleep icon. A new 'checkmark' icon image is used for skip. Otherwise this feature update adds a second button next to the unit skip button for the unit sleep action.