Migrate remaining docs, combine license files (#5044)
(1) Rename existing LICENSE file to LICENSE.md
(2) Migrate contents of license.md and add it as a header to LICENSE.md
(3) Update header text of license.md to remove external references to the license file as it is not included in the same file.
(4) Delete troubleshooting file in docs folder, last remaining file. The information is somewhat useful but so rarely needed that we can probably look it up again when helping a user debug a game freeze.