Fix double casualties being shown (#5064)
Bisected the error to refactoring update: 9c02b5dac2922c487b1acb05722756918ddfcf4e
The return value of `startGame`, an `IDisplay` is added to `IGame`. In the refactor, this add is dulicated and extracted leading to two calls to `IGame.addDisplay(display)`. This then doubles messages sent to the display leading to double casualties to be displayed.

The fix is to remove the extra `addDisplay` call.