Add hotkey (shift+d) for 'Done' Button (#5084)
* Add hotkey (shift+d) for 'Done' Button

Adds a key binding (shift+d) to execute the done action. Some minor refactoring is done to enable this. A tooltip is added to each done button to tell the user of the hotkey. Initially (ctrl+d) was chosen, but it had an OS conflict and did not work, hence 'shift' was chosen.

* Fix javadoc

* Formatting fix