Improvements to travis install script (#5115)
(1) Better commentary explaining http server launch from jar file is a hack
(2) Add -eu flag to fail early (best practice)
(3) Group variables closer together and add readonly flag
(4) Remove conditional on travis build variable. This is to support local builds. The previous commands to update DB do not succeed on locally to begin with. If supporting local builds, the full script should be updated and not just part of it.