Update git ignore to ignore image and sound asset file types (#5119)
Update game-headed gitignore to (for now) ignore txt placeholder files. Those likely can be deleted. Such placeholder files exist such that git will not remove an empty folder, that was likely needed as part of the release process to ensure we had all necessary folders present so that they could land in the release artifact.