Shellcheck travis scripts, fail early, no parallel (#5107)
(1) Remove parallel on build commands, it does not significantly improve travis build time (presumably the docker image travis spins up for us is single-core)

(2) Add a command to find all bash shell scripts in the '.travis' folder and shellcheck them. Shellcheck is a linting tool for shell scripts. Travis build fails if any scripts have a shellcheck violation.

(3) As early as possible in the build process, verify YAMLs, checkstyle, spotless (formatting) and run unit tests. This runs all tasks not depending on http-server or database such that the time to failure is decreased if we have any syntax, checkstyle, or unit test failures.

Some numbers for script runtimes (all numbers are for a fork build, does not include time to run release tasks):
- Clean build (master): 14m
- Clean build: 13m to 14m (about same)

- Checkstyle failure (master): 12m
- Checkstyle failure: 4m

- YAML syntax error (master): 7m
- YAML syntax error: 2m

- Compilation error (master): 3m
- Compilation error: 2m

- Unit test failure (master): 12m
- Unit test failure: 8m