Add condition to skip tagged builds (#5154)
In order to trigger a deployment, we push a tag to github. The tag creates what is effectively a branch, which travis then builds. This means we have master branch do a build, then as the last step it pushes a tag which then creates an addtional and unnecessary build.

Travis team seems to have provided this 'if' mechanism to skip tagged builds:
https://github.com/travis-ci/travis-ci/issues/8051#issuecomment-335664360

Benefit: Travis CI only provides so many VMs at a time, using VMs for needless builds wastes resources.

Side-note: the tagged build is the reason the build number jumps by two between releases. One build number is taken for the master build, building the tag increments it.