Remove out of date warning (fixes a deadlock) (#5211)
* Remove out of date warning (fixes a deadlock)

First, our 'out-of-date' warning code shows a dialog while we are in a background runner task. Swing components cannot be accessed while the background dialog is open. Because the background dialog runner is waiting for the dialog to be closed, we enter a deadlock situation.

The removal is valid because the dialog is actually questionable. If already check if the current engine is compatible with the save, here we are checking if the save was done with a future version. It should not actually matter if the save was with a future version because that version is still compatible. If there were a problem loading the save of a future version, it would not be a compatible.

* Run spotless, remove unused imports