Convert JPanelBuilder to a type-safe builder (#5257)
Calling a layout method on JPanelBuilder now returns a specific layout builder.
This allows for layout specific methods to be available only after invoking
that layout. For example, before a flowlayout could be used with a 'addNorth'
method, specific to border layout.

In addition, as a new feature, We add a new method to GridBagLayout to have an
'add(Component, GridbagConstraint)' method.