User-role table and add role 'HOST' (#5354)
With user-role being tracked in two tables, to avoid duplication of user role info,
a user_role table is being created.

HOST role is added to indicate a role that has no username or role id. This is the
role that is specific for 'lobby watcher' (a specific connection used to connect
to lobby to send game status updates).

To support further updates on serve side, AuthenticatedUser is being updated to
have username information.