Add ServerMessageEnvelope & Factory (#5384)
ServerMessageEnvelope is a unified java object that will be sent from server
to client over websocket. The ServerMessageEnvelope is designed to be JSON
encoded, it contains primarly two fields, a message type field and a payload
field. This is designed so clients can receive a ServerMessageEnvelope
as a JSON formatted String, clients can then convert it to a JSON object,
then based on the message type extract a java object from the message payload.

ServerMessageEnvelope objects should only be created by the corresponding
factory to ensure payload type is consistent and as expected.