Use Mockito timeout in websocket test instead of sleep (#5419)
Previously we had a manual sleep that was set to be as short as possible but 'long' enough
to still be reliable. Turns out the 10ms sleep was extremely optimistic.

We can instead use Mockito 'timeout' to give a longer wait window and make this test
more deterministic. With some manual test, test runtime is between 30ms and 80ms.