Allow unit-scroller keybindings only when unit-scroller is visible (#5456)
Unit scroller is visible/active when the move panel is visible.
This update checks if the move panel is visible, if so then the
keybindings for unit scroller are enabled. When the move panel
is not visible, the key binding actions are disabled.

An important note is that move panel is not rendered when its
not your turn and move panel is considered visible if it's in
a tabbed pane and is not the currently selected tab.

Motivation/Reason for change:
Keybindings are always active. Pressing 'n' key at any time will for
example highlight to the next unit, pressing 'space' will skip
and pressing 'space' many times will skip all units until
the 'no more units to move' dialog appears. This does not make
sense when it's not your turn or during for example a purchase phase.