DB updates, break up api_key table, add player-chat-id, system-id and consistency (#5438)
1. Break up api_key table into two:
  A) lobby_api_key: this table stores api keys for chatters
  B) game_hosting_api_key: this table stores api keys for
   'lobby-watcher', essentially the dedicated connection that
   hosted games create to send messages to lobby.
The reasoning behind this update is that for game_hosts, there
is very little information we keep track of. As one table, we
would have a lot of nullable columns. As two tables, we can
maintain non-null columns with better data expectations between
chatters and game-hosts.

2. Rename DB table 'username_bans' -> 'username_ban'
DB entities by convention are singular.

3. Add system_id and player_chat_id columns. System_id
is used to track the OS/system of a player beyond IP address,
player_chat_id is a field that will be used for identifying
players when banning.