Use URI for server instead of host+port combination (#5460)
* Use URI for server instead of host+port combination

This update changes the triplet "lobby.host", "lobby.port"
and "lobby.https.port" to be a single property "lobby.uri".
This drops configuration to connect to the old lobby and
allows us to define whether we will use HTTPS or HTTP
by configuration rather than hardcoding the protocol. This
is particularly useful as local development will likely continue
to be direct connection to (http) dropwizard while deployed contexts
will be to nginx fronting a dropwizard server over https.

** Configuration Updates and Consolidation of live versions file

'latest_version.properties' and 'lobby_server.yaml' are being combined
into one file. There are a few semantic changes to this configuration
file on top of the combined format:
- engine will treat version numbers as a "min-supported" instead of
  looking for exact version matches.
- error_message concept is replaced with 'inactive' flag. We had an
  error_message field before to allow us to notify users when they
  are connecting to an out-of-date lobby and that they should upgrade.
  The 'inactive' flag will instead trigger client side out-of-date
  messaging.

** Lobby Override URI Test-Setting Update

The engine setting for override URI is refreshed with this change.
The setting used to be a series of text field boxes and is replaced
with a radio button to select between a 'remote lobby' or 'local lobby'.
If the 'local lobby' radio button is selected, then a localhost URI
is auto-filled and used. Otherwise the 'remote lobby' has a free-form
text field for custom override values.

* Combine static initialization block

* Fix link

* Extract useOverride variable and use non-optional API

* Remove io.commons from test-common

* Use "Files" API instead of "FileOutputStream"

* Formatting

* Revert "Use "Files" API instead of "FileOutputStream""

This reverts commit 46e80d077b40799943dbb809d82aadd3b9d23f95.

* Remove out of date TODO

* Cleanup cast

* Add check that we have at least one active server configured

* Add double quotes around href

* Add @Integration annotation to integration test

* Use mock and fix test, clear cache

* Formatting