Improve connectivity chat error messages (#5551)
Shorten the error message and remove exception names from
the connection to chat error message and unable to send
message error message. Reduces noise in the error message
and more directly states the error.

Note, we still info log the exception so we can get the
stack trace in the error console.