Fix http_server role name & Invert deployment ordering of nginx and http_server (#5539)
Fix (revert) partial rename of 'http_server' to dropwizard, staying
with http_server for the moment.

Second, Nginx fronts http-server transparently, we need dropwizard to be
running for the nginx forwarding to work. Nginx depends on dropwizard,
not the other way round. This update inverts deployment ordering so that
dropwizard is deployed before nginx.