Implement User-Ban Filter (#5564)
Adds a filter to the dropwizard http-server to check if a user is
banned by IP or system-id. If so, their request is rejected.

- The main update is adding: BannedPlayerFilter and updating UserBanDao
  to provide a lookup method
- DB is updated to remove a constraint for ban expiries, enables test
  code to insert expired bans, and second we add an index for faster ban lookup

TODO: future iterations should consider caching a white list
and/or black list so that we are not sending a request to DB
on every http request.