Enable and test http-endpoint Roles (#5661)
* Enable and test http-endpoint Roles

1. We need a jersey config to activate allowed roles checking
2. Updated ProtectedEndpoint test to verify disallowed roles

The previous behavior was that if you could authenticate then you
could access any endpoint (@AllowedRoles was not active).

The prevous protectedEndpoint test verified that an admin
could access an endpoint and that an invalid API key (could
not authenticate) could not access an endpoint. This test
setup masked the problem that roles were not being enforced,
we only verified authentication.

This update updates the protected endpoint test to accept
a new enum 'AllowedUserRole' that specifies which role
can access an endpoint. To try and keep the test permutations
constrained, we'll verify that a user with that exact role
can access that endpoint, and second we'll test that a sample
of other (valid) roles cannot access that endpoint.