Remove guarded name constants (#5684)
'admin' was a special user name for chat connections in lobby. With the lobby re-write,
this should no longer be used. We can then drop all of the cosntants from LobbyConstants.

PlayerName and PlayerNameValidation converge once that check is removed as well, this
update replaces references from PlayerNameValidation to PlayerName.