Add config to restart http_server (spitfire) if changed (#5751)
Addresses: #5681 (Unable to Join Lobby)

We were running a stale version and had changed the server-client
data format causing errors on connection. Restarting the server
to pick up the latest version fixed the problem. The previous
deployment configuration did not restart the service, with this
update if we change the http-server jar file or server config
then we'll restart the service.