Remove rule-book links/buttons in favor of user-guide (#5952)
* Remove rule-book links/buttons in favor of user-guide

The user-guide page on website has links to the rule book, linking directly
to the rule-book is slightly redundant. Making the situation worse, the rule-book
PDF is becoming deprecated in favor of an HTML version.

This update removes links to the rule-book, everywhere we have a rule-book
link or button, we also had a link to the website 'user-guide' which in turn
has links to the latest rule-book.