Update PBEM validation to allow games without setting an email server. (#5981)
Enables the 'play' button on PBEM game even if email server is not
set. This allows for games with 'verified dice' using a dice server
and manual attachment of save games sent via email manually.

Previously it was required to also set email server settings which
allowed for automatic emails to be sent.