Fix account creation retry loop (#5989)
Account creation was written to have a recursive call for a prompt-create-retry
loop. On modification later, this loop became broken and is now just a hard
retry loop until server locks the user out for too many retries.

This update fixes this situation by aborting account creation if there is
an error rather than executing another retry to server. It is not
fully ideal as any user input from the account creation screen is lost
and the user must re-enter the create account flow from scratch.