Expire games (#6007)
* Notify game client when lobby listings are expired from cache.

Games are expired from cache if they are not renewed via keep-alive.
Currently nothing notifies clients when such a game is removed and
the client will think the game is still available. We previously
would have polled for all available games and the client would
determine when games have been removed. Then after switching to
websockets and an expiring cache, there was nothing done to replace
this notification mechanism.

Getting a listener to be notified on cache expiry is a bit tricky and
requires a cleanup job to be run. To extract these details, we
create a 'Cache' interface and consolidate CacheUtils to this interface.
Then we create an implementation that will invoke a removed item listener.
This is done by creating a scheduled thread that periodically calls
'cleanup' on the cache.

* Notify removal listener explicitly on 'invalidate' calls

This makes tests more deterministic and avoids delays for the callback
when waiting for 'cleanup'


* Use built-in Caffeine cache scheduler for maintenance operation scheduling

Maintenance operations, the things that trigger removals and removal listeners
to be notified are not reliably run all the time. The scheduler fixes this
so that maintenance happens more frequently and we get a dependendable call
to removal listeners.

* Explicitly define conflicting caffeine dependency brought in by errorprone