GameClient facade (#6060)
- Place most interactions to http client behind one of two facades:
 - PlayerToLobbyConnection: facades most of the interactions from player to lobby
 - GameToLobbyConnection: facades what was previously lobby watcher to lobby

The facade is partially a strangler pattern and will allow replacement of
http method calls with websocket. Secondarily, the facade is meant to simplify
network interaction. In the future more interactions will be websocket and
websockets will be combined on server side into two, one for player connections
and a second for connections from games.

Second, replaces most usages of GameDescription in LobbyGameTableModel
with direct usage of new data structure 'LobbyGameListing'. Eventually
GameDescription will be completely replaced with direct usage of
'LobbyGameListing'