Fix double "are you sure you do not want to move" dialog (#6183)
We were seeing two prompts of "are you sure you do not want
to move" when ending a move phase.

This is because of a call to "performDone" being done in both
a child class and parent class. This problem was introduced
by adding an override which then set the unit scroller to
invisible if we advanced. The same update inadvertently added
the double logic.

This update fixes that by relying on the parent method call
only and removing the override. The setting of the unit scroller
panel to be invisible is now done in the child class method
override that is invoked by the parent class 'performDone'.