Fix nullable accelerator to be uninitialized (#6215)
(1) Fixes menu accelerator used in a 'if not null' block, but
the value is a primitive and is never null. Hence, if an
accelerator is not specified, it was incorrectly defaulted
to zero rather than being a no-op. This fix updates
an 'int' field then to be an 'Integer' field.

(2) Add a parameter guide to fail early if we get a null
mnemonic, the value is expected to be non-null.