Make playerChatId non-null (#6237)
* Make playerChatId non-null

Assigns a player-chat-id to chatters that are connected to java RMI
games. This satisfies an assumption/desire for player chat id to always
be non-null. For java RMI games this is not immediately helpful, players
are still identified by their 'node' which contains IP address (security risk).

With player chat id, we have a unique, public, ID that can be used to
refer to players and could be used to replace 'node' as an identifier.

* Add player chat ids to test data