Battle display as dialog - remove always on top setting (#6251)
* BattleWindow from JFrame to JDialog

- Remove always on top property of the battle window
On some OS this seems to always win out over any other
notifications that are supposed to appear on top
of the battle window

- The conversion to JDialog will help ensure that the
battle window stays on top of the map and does not
get hidden behind it. This should have the same effect
as always set on top being set to true.

* Rename BattleFrame to BattleWindow as it is no longer a frame

* Remove PBEM.setFocusWindow

It appears to still work even with battle panel shown in a dialog
instead of in a frame.