Optimize findMaxLandMassSize() which can take 10% of AI time. (#6187)
* Optimize findMaxLandMassSize() which can take 10% of AI time.

This change optimized findMaxLandMassSize() by reducing the amount
of work needed to do. The main difference is we now keep track of
land territories already processed so that we don't need to start
new search from them.

This creates a new BreadthFirstSearch class that takes out the
BFS implementation I implemented earlier for finding capitals, so
that it can be re-used here.

It also fixes a correctness issue with finding the max land size -
the previous search would only search up to a distance of 6. The
new implementation does not have this limitation.

On Domination 1914 map, which has a lot of territories, this saves
about 50 seconds of CPU time over a turn of an all-AI game.

* Optimize findMaxLandMassSize() which can take 10% of AI time.

This change optimized findMaxLandMassSize() by reducing the amount
of work needed to do. The main difference is we now keep track of
land territories already processed so that we don't need to start
new search from them.

This creates a new BreadthFirstSearch class that takes out the
BFS implementation I implemented earlier for finding capitals, so
that it can be re-used here.

It also fixes a correctness issue with finding the max land size -
the previous search would only search up to a distance of 6. The
new implementation does not have this limitation.

On Domination 1914 map, which has a lot of territories, this saves
about 50 seconds of CPU time over a turn of an all-AI game.