Suppress lobby unnecessary player joined message & premature player left message (#6395)
* Fix double 'player joined' message when a registered player double connects

Registered players are allowed to have multiple sessions from different
game instances and would chat as if under one name. On the second connection,
it is confusing if a 'player joined' message appears even though the player
is already present. This update suppresses that message.

* Only broadcast player has left message when the last session has disconnected

This update only sends a 'player has left' message when the last
of potentially multiple sessions has disconnected.

Registered users are able to connect multiple times under the same
exact name (in 1.9 a player would have '(1)' suffixed, in 2.0 the
multiple connections is transparent). In such cases, when one
session leaves, we do not want to broadcast a player left message
as the player is still actually connected (which is weird when the
player chats but is not listed in the chatter list).