Fix unit factory toctou (#6385)
* Simplify, avoid assigning to Optional type, instead use 'map' operation

* Fix TOCTOU when getting images from an image cache.

There exists a 'clear cache' method that can possibly remove
the data in 'images'. If a method call interleaving is just right
then we have a TOCTOU problem where we can detect an image exists,
clear it, then try to retrieve it resulting in a NPE.

This is fixed by getting the image and checking if that is null,
if not we use the image we retrieved.

* Remove the need to clear image cache in UnitImageFactory by using a new instance

- Change 'setResourceLoader' to instead to be a constructor. If we use UnitImageFactory
before we have called 'setResourceLoader', we would get a NPE. This implies that it
is safe to convert 'setResourceLoader' to be a constructor.

- When changing scale factor, instead return a new UnitImageFactory that has a fresh
cache.

With the two changes above we no longer need to clear cache.