Fix lobby chat messages not being sent with 'enter' (#6405)
This update fixes it so that pressing enter once again sends chat messages.
The problem was after consolidating all key bindings to use the same utility
class, that utility class suppressed all key actions when a text component
was active (to prevent for example from the 'f' key firing its binding
while a person is simply typing). As part of collateral damage from
this is bindings to the 'enter' key. This update simply creates
an exception for the 'enter' key so if a text component is in focus,
keybindings to the enter key will fire.

Addresses: https://github.com/triplea-game/triplea/issues/6353