Moderator tab remove button col headers & rename super-mod -> admin (#6492)
* Update Toolbox:ModeratorsTab - remove button column headers

Remove column headers "Remove Mod" and "Make Admin". The column
headers are redundant for button columns where buttons appear
in every row with the same name as the column header.

* Rename 'Super-Mod' to 'Admin'

Historically a moderator was called an admin. In 2.0 with
user-roles, the 'admin' role is now called a 'moderator'.
To avoid previous confusion the term 'super-mod' was used
to refer to an admin. This update replaces the 'super-mod'
term with 'admin' (to recall, an admin is someone that
can add and remove moderators).