Do reverse-connection check of new game hosts when posting games (#6519)
This update combines the connectivity check endpoint with the post-game
endpoint. This avoids any timing issues and seeks to simplify
the game posting flow. Instead of a client doing two round trips
to the server to first post a game and then ask for a connectivity
check, it is done as one now.

For implementation updates, we update the response from the game-posting
request to include both game-id (nullable) and the result of the
connectivity check. After posting a game we now should check
the connectivity result and if okay, then proceed, otherwise
execute the previous behavior when the connectivity check failed
(which is to show the "this is how to set up port forwarding" message
and shutdown).