Fix NPE in auto-placement-finder when finding islands (#6580)
* Fix NPE in auto-placement-finder when finding islands

The contract of MapData#getContainedTerritoryPolygons (in javadoc)
states the parameter should be a sea territory. The caller passes
all territories to this method which then results in a NPE in
the method. To fix, relax the contract and return an empty collection
if the parameter is not a sea territory.

Reported in: https://github.com/triplea-game/triplea/issues/6469