Infrastructure: deploy admin scripts to http-server (#6622)
When logging in to http-server as admin, this update
deploys utility scripts to /home/admin, like 'server-start.sh'
to start the server, 'server-logs.sh' to see logs. This is to
make it very easy to know which commands need to be run
in order to start/stop the server and view logs.