Fix npe with out of date version (#6711)
* Fix NPE triggered with out of date client versions

Inactive server config does not have a host-uri. Yet, we marked the server
URI as non-null when creating the server properties object and also
when we create a LobbyClient object. This resulted in a potential NPE
if a client version sees an 'inactive' server config in servers.yml.

This fix removes a non-null constraint as server URI can be null,
and moves the handling of inactive servers to be before we create
a lobby client (avoiding the second potential NPE).

Co-authored-by: tripleabuilderbot <tripleabuilderbot@gmail.com>