De dupe error uploads & Add Map Name to Error Uploads (#6984)
* De-Dupe Error Report Uploads

Prevent redundant error message uploads by first checking
with the server if an error report already exists.
If the error report exists, then we will show a prompt
to the user and give them a link to the existing error report.

Change Summary:
- Add columns in error report history table to track the
  title of the error report, reporting game version,
  and the issue link that was created.
- Add APIs to error reporting DAO to get the link of
  an existing error report by title and game version.
- Add server API to check if an error report exists
- Update client so that after clicking 'upload error report',
  we first query the server if an error report exists.
  If one exists, we show the link to the user, othewise
  we show the user the upload form.

Note: we use both title and game version for de-duplication.
  This may lead to some duplication as version number
  changes, but in case we do not fix a problem or if
  the title matches but is actually different in a different
  version we will allow the error report upload.

* Add field to error report upload for map name

* Do not print Exception in its own block (redundant)

Exception is always at the head of the stack trace which
we already print in the issue body. This update removes
the redundant 'exception' block.