Hide deployment variable output on failure - fix leaked password (#7287)
* Add no_log = true to flyway database deployment tasks to prevent password leakage

Tasks that fail with loop print out the loop item. The loop item for database
deployments contains a password and is thus printed out if there are any failures.
This update adds a 'no_log = true' option which would suppress this output.

* Rotate maps_db password

The password was leaked in travis logs in plaintext after deployment
failures. The failure output printed the loop variable that contained
the password.