Speed up db rider tests (#7488)
* Update DropWizardExtension: clean after each test & init JDBI once & fail if server already started

1. Re-creating DB connection takes some time, if we already have a DB
connection (ie: jdbi != null), then do not re-created it. This is
a matter of adding an if null check around the JDBI instantiation.

2. Clean DB after each test, this helps avoid FK violations between
tests within a test class. The execution of the clean-up is quite
fast, this makes it easier to write valid tests without slowing
things down noticably.

3. To improve error handling and error messaging, configure the drop wizard
extension server startup (starts server when tests execute) to fail if
the server is already running. Given we no longer manually start a server
in CI, this is okay.

* Speed up Tests: Update all @DataSet to have 'useSequenceFiltering = false'

This is a big boost in performance for DBRider tests that has more than a
few tables. SequenceFiltering checks the right order in which to insert tables,
presumably by attempting permutations. By not doing this, we see pretty significant
test execution speed-ups.

A downside to this udate is that DbRider will typically fail if we define
multiple tables in a single YML. The order of tables in YML does not determine
the order of creation in DbRider. To fix this, we can put each database table
in its own YML and include multiple YML files per @Dataset. The order of included
YMLs is deterministic relative to table creation and in this way we can
ensure tables are created in the right order.