Add xml-reader subproject, a StAX based xml-reading library (#7543)
StAX is a DOM-less XML parser. StAX is essentially a cursor that
traverses an XML document and notifies the code of parsing events
as they happen.

This update adds a subproject, 'xml-reader' that wraps this library
and provides an annotation framework (very similar to JaxB) where
Java objects that model the XML can be annotated and automatically
created by 'XmlMapper'.

The sub-project README has details and documentation on how to set
this up.

As a feature the mapper can automatically map 'int', 'double'
and 'boolean' objects.

Of note, the StaX parser does not validate XML against a DTD.
If we want that validation we'll have to do so as an auxilary step
(perhaps something we can wire up for map-makers to validate
their XMLs).

Overall the sub-project is designed to give us a lot of control
over XML parsing. For example a logical next step will be to
add 'Aliases' to the annotations to support different naming of
equivalent tags and attributes.

The xml-reader is also designed to be efficient (fast) by using
StAX and emphasis has been put on yielding good error messages
when things go wrong.

As a larger fit, the XML-reader will be used to create data
model objects of the Game XMLs. These data model objects
will represent the data in XML almost exactly and then
will be used in another stage for 'game data assembly' where
they are put together to create a game data object.
This is different from 'GameParser' of today that parses
XML and directly creates a GameData object. This extra
layer should allow a lot of flexibility for different
XML formats, notably to support both future and legacy types.