Reduce state in GameChooser & Apply Strangler to Shallow Parsing (#7552)
1. Stop keeping tracking of a fully parsed vs shallow parsed game data in game chooser.
  Instead we'll return a fully parsed game data when requested and otherwise
  we keep track of the shallow parsed data. It turns out the shallow parsed data is
  just the game name.

2. Extract shallow parsing to its own module so that later we can swap out
  the implementation (strangler pattern). As part of this we can simplify the API
  of shallow parsing since we only need the game data name.

3. Reduce severity level for invalid maps from 'SEVERE' to 'WARN'. Bad map
  data we do want reported to users but they are not actionable bug reports
  for us generally (users who are working on new maps would expect to see problems).

This update changes error handling behavior such that if a map has a parse error we do
not deselect it from the game chooser. This is a bit forward looking as eventually
a shallow parse will not reveal any errors anyways. It would make sense to get
a deep parse only when selecting a game map and seeing errors then.

This update does reduce game performance in the map selection as we do a game parse
whenever clicking any new game selection items. We'll need to revisit this to
see if and how that can and/or should be improved later.