Add a band-aid fix to avoid a ConcurrentModificationException (#7598)
Likely a better fix would be to ensure that any UI models
that need to 'hear' events have finished registering before
we could fire any. It appears that during a game load such
a sequence occurred where an event was fired while something
was still registering a listener resulting in an error.

This fix is a band-aid as doing a list copy is not ideal
long term. It is better though than errors and it might be
some time before the game startup sequencing is clear enough
that we can avoid the noted scenario.

Related to: https://github.com/triplea-game/triplea/issues/7588