Remove 'isBombing' boolean, instead infer it from BattleType (#7643)
* Remove 'isBombing' boolean, instead infer it from BattleType

This update adds flags to 'BattleType' that can be used to directly
infer if a battle type is bombing or an air battle. This has a
relation to the 'isBombing' flag that is passed around into methods,
so rather than using methods like "boolean isBombing, BattleType type",
we can replace that with "type.isBombing, type".