Structure lobby login window so that it can be re-used (#7749)
To be able to do a lobby login prompt from other screens,
we need to decouple the successful login action from the
login screen.

This update restructures responsibilities so that the
caller of the login screen can inject the successful
login behavior.

Additionally, this update adds a flag that can disable
unregistered login. The reasoning is if we are requiring
a login to enable a functionality, then we need
the user to have or to create an account.