Fix: Use all of the friendly units for support calculation in OOL (#8000)
If the targetsToPickFrom is a subgroup of all of the units on the other
side, then the support calculation will be incorrect.  This could cause
the OOL to be calculated incorrectly.

An example would be submarines fighting air planes and other surface
ships.  The submarines can't hit the air planes so the targetsToPickFrom
is just the surface ships.  But the air planes might give support to the
surface ships and make some ships more valuable than normal.