Hide usages of websocket 'Session' being an interface (#8063)
Preps for a potential standalone websocket server to replace java sockets.
The standalone server has a different object type than a session, by
hiding the 'session' behind an interface we can accomodate for this.

A more significant change to this update is we will still always
use a 'GenericWebsocket' for server connections, but instead of
embedding fields into a 'session', into 'user properties', instead
we'll create a static map of classes and their respective generic
websocket handler. Then, the websocket server classes will fetch
the generic websocket from this map. To initialite this, we
now use an 'init' method to bind a websocket class to its
generic websocket server.