Make the gameChooser use DefaultGameChooserEntry instead of String (#8080)
* Make the gameChooser use DefaultGameEntry instead of Strings

In a map, a designer might accidentally include two map files with the
same name.  Both will show up in the UI but because the UI uses the
String as the key, selecting either of the two entries in the UI will
always load the same map. By changing the UI to use the
DefaultGameEntry object, both of the UI items will select their
corresponding entry.