Reword "units to place" text -> "purchased units" (#8222)
"Units to place" is confusing as the verb tense makes it
sound like it is an immediate action occuring during the
current phase. Considering the 'placement units panel'
appears in all phases except the placements, it's confusing.

Even worse, the 'units to place' text appears next to
'units to move', the latter is an immediate action that
is occurring during the current phase, the former is a
future action (confusing).

"Purchased Units" is direct and is usually semantically correct
on most games (arguably in others, the purchase price was zero
for some units, hence still semantically correct).