Remove code to 'warm up crypto random source' (#8221)
The commentary code states warming up the crypto random
source takes a long time. Measuring the runtime on a bot
and local hosted game, it takes either nano seconds
or runs under 100ms.

We have errors in the logs about this action, and it
does not seem to clearly meet the stated performance
objective as the invocation is quite fast.