Use a value object to combine the rolls and strength in power calculation (#8250)
Rolls and strength values are intertwined with each other when dealing
with power calculation.  If one is zero, then the other is zero. Instead
of keeping track of these details in the callers, the StrengthAndRolls
value object will ensure it.

Also, pass around the RollValue and StrengthValue as much as possible so
that it can be type checked and the information they store can be kept
through changes.