Allow validating canals when finding neighbors or distances (#8289)
* Allow validating canals when finding neighbors or distances

The ProAI calculates the value of territories based on neighboring
territories and distances to enemy factories/capitals. If a neighboring
territory had a canal to it, the ProAI would ignore the canal and assume
that the units could travel to it. This could cause unit stacking
because of the higher value of the territory.

Now, when the ProAI calculates the neighboring territories, it checks if
units can travel between between the two territories. This will cause
some neighboring territories to have a higher distance and so the value
of the territory will go down.

* Fix unittests matchers

* Create another two overloaded getNeighbors methods

* Add overloaded getDistance method

* Inline Matches.alwaysBi instead

* Change variable names and lengthen comments

* Add shim code to combine predicate and bipredicates to be a single call to a bipredicate only method

* Remove shimmed methods, change invocations of (predicate, bipredicate) methods to invoke (bipredicate) methods directly

* Remove another shimmed method

* Fix refactoring mistake, use correct variable in condition.

Co-authored-by: DanVanAtta <dhvatta@gmail.com>