Rework the auto low luck casualty selection (#8338)
* Rework the auto low luck casualty selection

Rewrite the old comments to fit better with how the game now works. When
this was first worked on, all AA were infinite, the dice sides was
always 6, and AA only had a strength of 1 or 2.

The new logic is a little different from the old logic. The old logic
used a little more randomness than the new logic.  But that randomness
doesn't appear to improve the functionality so it was removed.

* Make LowLuckTargetGroups a top class

Move specific logic of LowLuckTargetGroups into it.

Call buildCasualtyDetails in getAaCasualties instead of each of the
called methods.  Change those methods to return the hit units instead.

* Set the player name for DiceRoll to be an empty string in the test