Headless logging conversion to logback (#8466)
* Add headless logback.xml & convert 'chatHandler' to be a 'chatAppender'

This changes the headless server handling of logging java.util
logging to chat to instead be slf4j based and to send
slf4j logging to chat. As part of this change we also configure
headless logging to have a logback.xml which configures logging.

Functionally this does change the output and there is no longer
a stack trace. This is generally okay, we have yet to receive
any bug reports where the output message from a bot into
chat was reported as a bug (generally these logging events
are treated as informational and otherwise ignored by players).

* Rename ChatHandler -> ChatAppender

* Simplify Appender and remove thread safety guards

J.U.L handlers are not thread safe, logback 'AppenderBase' is
thread safe.