Remove check to skip downloading already installed maps, this breaks map updates (#8494)
When we instantiate a download window with a list of maps to install, these could
be maps that we wish to update. They will be detected as installed, there is logic
that will skip downloading already installed maps, which breaks this flow.

To fix the update map flow, this check for maps that are already installed is removed.