Increase file line count and cognitive complexity thresholds (#8495)
This update sets the threshold values for line count and cognitive
complexity to be relatively extreme values. This is to avoid
false positives with either metric such that if the metric fails
we can really pay attention to it and have it flag the cases where
we really would want a fix.