Remove wip map upload (#8651)
* Update map design doc to reflect decision to keep maps in github repositories

Reflects the results form these two conversations:
- https://forums.triplea-game.org/topic/2165/storing-map-index-in-database-deprecating-triplea_maps-yaml
- https://forums.triplea-game.org/topic/2450/rfc-in-game-map-uploads

It was debated whether github repositories were a good place to store maps and if we should
have an in-game UI to uplodate maps. It was decided that github repositories provide a lot
of features for us and would not be replaced by a custom solution.

* Delete WIP code to support map uploads

We will not be uploading maps via the in-game UI, the plan is to continue
having maps uploaded into github repositories directly.