Fix loading map names with a numeric name (#8774)
Fixes loading maps with names like '1941'.

When stored in game data properties, the value at some point
turns into an Integer. When extracting we attempt to cast this
to a 'String' and this then causes a ClassCastException. To fix
this, we can instead do a 'String.valueOf' rather than do an
explicit cast.