Fix bad-zip folder could not be created false-positive error message (#8772)
If the 'bad-zip' folder already exists, the game engine when trying
to move bad-zips into that folder would give a false-positive
error message stating the folder could not be created.

The cause behind this is 'mkdirs' returns false if a folder already
exists. This update fixes that problem by checking if the bad-zip
folder exists, and if so skips trying to create it.