Update triplea_maps.yaml (#8839)
Added new pull request as missing the img the first time. 

I don't really understand why it needs the \blob in the address as the others don't seem to, but if I go and open the image from the repo that's the address it uses.  Also doesn't seem successful without it.