Use the default selected casualties when a single unit type is targeted (#8858)
The allTargetsOneType code assumes that all the targets are identical.
But with support and other new aspects, two units with the same unit
type could have different power. So use the default casualy order of
loss calculation for all target situations, but don't ask the user to
pick the casualties since the dialog is redundant.