Non-Modal Combat Prompts (Allow user to scroll map during bombard prompt) (#8949)
In short, this allows users to scroll the map during combat prompts such
as bombardment (attack subs, etc)

This update converts mainly combat confirmation dialog prompts to be non-modal
and uses a latch to simulate the blocking nature of a modal prompt. The benefit
of a non-modal prompt is that a user can scroll around on the map. This is very
handy for example when there is a territory bombard prompt for multiple territories
that are offscreen.

In addition, this update improve the API for these confirmation prompts:
  - return a simple true/false to indicate if user confirms or not rather
    than a magic number.
  - use an enum input parameter instead of magic number for confirmation type.