Fix confirmation dialog deadlocking EDT when launched from EDT (#9008)
The non-blocking confirmation dialog recentl introduced blocks
the current thread while waiting for a user to select a confirmation
dialog prompt. If the current thread is the EDT, then the dialog
prompt that needs to render on the EDT also becomes blocked.
This creates a deadlock. To resolve, if the confirmation dialog
is created on the EDT, then we make it modal and blocking,
othewise if we create it off of the EDT, then we block the current
thread and the dialog will be non-modal.