Auto-select battle calc sides when using 'ctrl+A' and 'ctrl+D' (#9012)
This update adds two behaviors for the battle calculator when using
the hotkeys 'ctrl+A' and 'ctrl+D' to add attackers or defenders
to reduce having to switch the current attacker or defender
in the battle calc.

When adding attackers, if there are no matching attackers with
the current side, and there is a defender present in the battle
calc, then the attacking side is automatically set to the player
whith the most units in that territory. For example, if a player
has set up 1 defending UK infantry, and there is a territory
with 1 german tank, and the battle calc has italy set as the
attacker, then using control-a will automatically add the tank
and set the attacker to germany.

For defenders, if there are no units added to the battle calcuator
at all, then if a player uses 'ctrl+d', the defending side will
be set to match any unit that is in the current territory. For
example if a default calculator has germany as the defender, and
a player uses 'ctrl+d' in a russian territory with russian units,
then those units will be added to defenders and the defending
side will be automatically set to russia.