Use a simpler content downloader on maps server (#9051)
Allows for status code to be received by caller and then processed.
Notable as before a 404 triggered an error logging from within the
library with stack trace. For cases where a 404 is expected, this
is not desirable behavior. This update adds a simpler version
of the downloader library that throws an exception that provides
access to returned status code. The net result is the logging
for 404 can be handled by client and print a more friendly and
expected message rather than a scary error message with stack trace.