Do not set invalid 'test' auth token for github API (#9308)
By default the auth token set to use with github is 'test'.
Sending this to github results in an 'access-denied'. This is
good if smoke tests are running, otherwise is not desirable.
Sending no token is fine, it incurs a rate limit.

This update does not set the github auth token when set to
the default value. This allows map indexing and other github
API interactions to work out of the box.