Improve SpitFire server test configurations (#9375)
Renaming updates and documentation enhancments to improve clarity
on spitfire server test configuration modes.

Notable, we no longer use a magic github API token of 'test',
it is now either left blank or specified. Instead a new feature
flag is used in the place of checking for an API token of 'test'.

A new consistency is put in place here where out of the box
features are turned off, but all turned on by default in
prerelease and production.