Update stalebot config: only issues and older (#9467)
- Updates stalebot config to consider issues as stale if they
180 days old and only close if they are open for multiple years (5)

- Improve stale bot 'stale' commentary to ask for a summarizing comment.

- Only mark issues as stale, no longer PRs.

Motivation: PRs marked as stale is not very helpful. The previous
stale comment was not that helpful for getting issues resolved,
it's better to ask for an issue to be summarized rather than
re-partitioned and re-tasked as was previously commented. Second,
the 90 day window to close was actually pretty short, particularly
for a project that is not worked on consistently all the time, hence
issues were being closed a bit too promptly.