Update ansible infrastructure to only deploy a single DB and show flyway migration results (#9499)
We had to 'no-log' the flyway migration results to avoid leaking the database
password. A downside to this is if there were any migration errors, they would
not show up and further troubleshooting would have to be done.

The database password was leaked because we were looping over a list of databases
and ansible will print out the contents of loop variables.

This update changes database configurations from being multiple to just allowing
only one database. This then allows us to show the flyway logging without
leaking a password. Meanwhile, we are likely to stay on the one database for
some time.