DropWizard - Use Forwarded Headers (#9500)
In DropWizard 2.x the default to use forwarded headers was turned off.
This update turns that configuration back on. The configuration is important
for when behind a reverse proxy. When behind a reverse proxy, if we don't
use forwarded headers then 'getRemoteAddr()' will return the IP of the
reverse proxy and not of the proxied client.