Make CTRL-A, CTRL-D, CTRL-B work with focused battle calculator dialog (#9518)
* BattleCalculator: Make CTRL-A, CTRL-D work

The key bindings to add attackers or defenders, namely CTRL-A and
CTRL-D, only work if the main TripleA frame has the focus.

Add the same key bindings to the battle calculator dialog, so they also
work if a battle calculator has the focus.

* Battle Calculator Make shortcut CTRL-B work

So far the shortcut CTRL-B to open a new battle calculator only works in the
main TripleA window.  Add this shortcut also to the battle calculator dialog.

Note: The new battle calculator is opened for the currentTerritory which is the
territory the mouse was over last in the main TripleA window.  Mouse movements
in dialogs are not seen by the main window and thus do not update the current
territory.