Battle Calculator: Allow fighting any player (#9532)
* Battle Calculator: Allow fighting any player

When changing a player, do not set the other player automatically to an
enemy.  This allows fights between any two players.  Moreover, it does
not erase the battle configuration in case of a misclick.

* Battle Calculator: Mimic legacy behaviour for intial setup

Mimic the legacy behaviour and force the two parties in the initial
setup to be non-allied.  This legacy behaviour was requested in #9532.