Maps/out date detection by modified timestamp (#9576)
* Update out-of-date detection to use file timestamp

(1) Change detection of out of date maps to use file timestamps.
If the last commit date of an available map download is before
the last modified date of a map's install folder, then we will
the map on the file system to be out of date.

(2) Maps with an update available are now also shown in the
download maps window 'installed'. Before they were *only*
shown on the 'out-of-date' tab. By showing 'out-of-date' maps
on the installed tab, a user can uninstall the map. The only
option for a map on the 'out-of-date' tab was to update it.