Refactor: Consolidate ownership of  'skin name' -> 'skin path' mapping to 'UiContext' (remove from UI) (#9770)
UI layer previous had a lot of knowledge about UiContext (effectively a model):

    knows the definition of the current skin
    knows skins names are mapped to skin paths
    knows that UiContext uses skin paths to set map directory

Instead we want the UI to get a list of skin names and be told which one is selected. The UI ought to then render this and then inform the model when a new skin name has been selected.

This PR would be best reviewed commit-by-commit or started by reviewing the UI changes first.