Fix for 'unable to move properties file' following map zip extraction (#9819)
Not sure why the code was trying to move a properties file following
map zip extraction. Instead we should delete these files as they
are no longer used.

Meanwhile, the presence of any .properties file can generate
an error message on startup. Notably this can be done if we re-download
an already extracted map and create a new .properties file which
would then collide with the already moved file.