Updates for additive map skins (still work in progress) / Put map skins behind a beta feature flag (#9821)
An incomplete update to try and get map skins working again. Notable changes:

    map skin functionality is placed behind the beta feature flag
    map skins are now additive to the base map. Meaning if a map skin is missing an image, we'll default to the original map and then to the game engine.
    map skins are detected via the presence of a 'skin.yml' file and are located within a map
    missing unit image warning message will now print out search paths
    removed an invocation to 'parse all maps' when rendering territory effects, this could improve performance

Next Steps:

    ResourceLoader when changed mid-game does not update cleanly. This needs to be fixed up for map skins to be brought back out from the 'beta' feature flag.

Change Details:

* Refactor: replace 'UiContext.getResourceLoader.getMapLocation' with 'UiContext.getMapLocation'

* Add a 'skin' path parameter to ResourceLoader & update resource loader ordering

If present, we will load from 'skin path' now first before we load from map path
or the game engines path.

* Add Util 'PathUtils', with method: #toUrl(Path)

* Refactor: ClipPlayer/use an explicit 'sounds' directory resource loader

This avoids the need to have a nullable ResourceLoader param

* Refactor: move construction concerns to UiContext

* Docs: add javadoc to method ResourceLoader#findDirectory

We should move this method to 'FileUtils'. In the meantime added documentation
to clarify that the 'findDirectory' starts from a given path and then 'up'
to parent folders rather than child folders.

* Refactor: (Cleanup from previous) remove unused 'mapLocation' variable in resourceLoader

* Refactor: ResourceLoader constructor, simplify to be just a list of Path that are the priority search order.

Though, as a comment states, the search order is not well respected. Fixing this is left as a TODO.

* Refactor UiContext to use a list of resource loader paths

Removes conditional logic around 'setInternalMapDir'. Before
the map directory was considered to either be the skin
directory or the map directory. Now the skin directory
is just an optional first element in the list of resource
directoriers (with game engien assets directory being
last in that list by default).

* Refactor/Cleanup: reorder lines

* (Possible performance gain): inject resource loader into map data and re-use existing resource loader for territory effect images

* Refactor: rewrite #getTerritoryEffectimage to use 'Map#computeIfAbsent'

* Refactoring fix: update usage of 'ResourceLoader' constructor

* Refactor: Convert MapImage 'loadMaps' to be a constructor

* Cleanup: fix IDEA warnings, redundant initialization (initializers that have no effect, ie: dead-stores)

* Partial implement 'findMapSkin'

* Refactor: convert 'yamlFileLocation' type to 'Path' and add clarifying comment

It was previously confusing if the 'URI" property was the URI where the data is
avaialable for download or if it was a path on file system. By converting
the type to a path it is more clear and simplifies.

* Implement 'find map skin'

- Refactor:   Rename 'find' to 'findOne', and add a 'find' method to return a list of file results.

* Various fixes for skin functionality

* Put 'Map Skins' behind 'beta feature flag'