Wire map download length from server to client (#9854)
We are already capturing download length in database. This update reads
the file size download value from database and updates the server
response for map download listings to have the download length. In turn
we update the client to use this value (and remove code that previously
did a 'head' request as an attempt to get download length)