Simplify/Update Error message formatting (#9931)
Update error message formatter to print logged text or else print 'an unexpected error ocurred'.
We then print the top 3 exceptions and their messages.

- We are able to greatly simplify by virtue that 'loggerRecord.getMessage()' is non-null.
Though, if we thrown an exception, then the message of the 'loggerRecord' and that of
the exception will be the same. This leads to duplicate printing of the same message.
This is why we print 'an unexpected error ocurred', in such a case there is an exception
and we will print its message on the next line.