Fix selecting step nodes in history view not having an effect. (#10439)
Fix selecting step nodes in history view not having an effect.

This was broken with #10429, which made the logically correct change, but ran into a bug where we weren't actually setting the change index on step nodes. Introduce a fix as well as handling of history from saved games that didn't set it.

Additionally, when non-leaf nodes are selected from the UI, we actually don't want to go to the state after that node, since that breaks chronological property of elements of the tree (i.e. Round 1 non-leaf node would be before first leaf node of Round 1), so this PR adds logic to go to previous child node instead.