Make ChangeAttachmentChanges save a copy of the previous value. (#10483)
Make ChangeAttachmentChanges save a copy of the previous value of the property.

The previous value is used when inverting the change (e.g. when going back in history and reverting changes). By copying the properties, this fixes several issues:

* Ensuring that the inverted change is actually correct, as we don't want to be reverting to a "live view" of the current state.
* Fixing exceptions caused by my recent change where property getters return unmodifiable collections (we don't want to set the property to an unmodifiable collection, since future changes will fail).

Additionally, switches to improved syntax for some locking in a few related places and some small clean ups.