Add a units bar, showing units inside the territory after its name in the bottom bar. (#10404)
Add a units bar to the main UI, showing units inside the territory after its name at the bottom of the screen.

This feature makes it easy to see which units exist in the given territory, without having to open the battle calculator or switching to the territory tab (which loses focus on that territory when you move the mouse). This is especially helpful on some maps where there are small territories that can get cluttered with units and it's hard to see exactly what's there.

This change makes it so the units in the territory under the cursor are shown at the bottom of the screen, in the same cell in the bottom bar where the territory name is.

I've taken special care to test that this works well on a variety of maps and results in consistently good looking results.

In particular:
  - The content of the cell continues to be centered.
  - On smaller screens, priority is given to the territory name / resources if not everything can fit.
  - It works well with a variety of image sizes and presence of larger images doesn't cause layout issues.
  - A menu item is added to toggle the display of this bar, so users who don't like it can turn it off.

Tested manually resizing window to smaller and larger sizes to verify layouts on different screen sizes.