Add an end-to-end test that executes an all-AI game to completion. (#10522)
Add an end-to-end test that executes an all-AI (hard AI) game to completion.

Currently, only doing this on the very simple "mapmaking tutorial map", which exercises sea zones and amphib invasions. In my testing, this takes 1 min to run and a game completes in 24 rounds.

Also adds a few more maps to GameSaveTest.