Infra: remove configuration to deploy to prerelease (#10708)
We will either be deploying to a local 'vagrant' test environment,
or production. This update removes the 'prerelease' configuration.
Since we will also be releasing production from 'tip', we can keep
the code that builds latest artifacts and deploys those artifacts
rather than downloading artifacts from github releases. The main
advantage to this is we get better unity between test and production
deployments.