Update verify script (#10711)
* Dev-UX: Update verify script to check for docker dependency

We modify the verify script to look for 'docker' so that we can
fail early with a more helpful error message. Future improvements
might try to auto-install docker or at least provide a link to
install in the error output.

Now that the verify script is more than just a couple of commands,
We also change the structure of the 'verify' script to have a 'main'
block.

* Simplify & Fix: Properly skip database tests before database has been started

The verify script has a command to start database, but this is done only after
we have run 'check' in parallel on all sub-modules except for those that
depend on tests. The database dependent tests were specified down to the exact
sub-module, which did not scale as we added a sub-module but did not update
this white list.

The result of having a test start before database is that we got an error message
that the database was not available. This gave the false impression that we
need to start database before running 'verify' (when really it was just broken).

This update resolves this by specifing the lobby top level module that we will
skip without specifying each individual sub-module below the 'lobby top level'.

To better root out any problems with database starting up, we move that to be
started first before we execute any checks.