Teach Hard AI to save up for a fleet. (#10675)
With this change, Hard AI will not use its unused PUs to buy high-value non-sea units (that otherwise aren't needed) if it needs to save up for a fleet. It will do this if there are no enemy territories reachable by land from its factories and it's able to build fleets that can reach enemy land. Additionally, the logic only triggers if we tried and failed (due to not being able to defend) to place sea units and only if we're not already at the resources needed to build the max fleet we can.