Infra: run certbot 'certonly' to avoid updating nginx configuration (#10753)
Certbot adds a new line to nginx config after it runs, even if no
other changes are made. This causes some trouble in automatically
restarting nginx. We hack around this by running a nginx role
after the certbot role to restore the configuration file.

If we did not do this confguration restore, then the next lobby
deployment would cause a change in configuration (removing the new
blank new line) which then woudl trigger a needless nginx
restart.

There is a flag on certbot to only get a new certificate, which
is useful for us here so we don't have to do special work restore
the nginx configuration.