ProAI: Improve territoryHasLocalNavalSuperiority(). (#10797)
In addition to estimating strength differences, check that we also have a positive TUV to attack the strongest defensive enemy flee that's nearby.
This prevents a situation where the AI does not expand its fleet yet does not attack due to its fleet not being strong enough.

An example of this is on the save here on: #10785, where this situation causes a stalemate. With this fix, I confirmed that the above game gets out of the stalemate with the expected result of the AI with more production eventually winning.