Fix .github workflow do-release script (#10803)
BUILD_NUMBER variable was empty and this caused problems later
when we had an empty 'tag' name for the github release action.
Previously the 'set-build-number' script returned the version
number. That behavior was updated and hence it is now empty.

To correct this, we make use of the new scripts that can return
the build number (which do so in an idempotent and deterministic
way, without mutating other files as a side-effect).