Infra/versionless server artifacts (#10871)
* Only set version number on game-client artifact

Removes the version number on jar files except
for the game-client jar file

* Simplify build artifacts script

We no longer need to do an explicit clean to remove prevoiusly built artifact.
We had that in place because the previous artifact might have looked the same
to gradle (and hence not be built, because not changed), but could have had
a different name. Now that version number is no longer in the name, we are
guaranteed to find artifacts built by gradle.

* Remove version number from artifact names

* Add a version.txt file to lobby deployments and de-dupe deployments on major.minor

- The version.txt file is to help us know which exact version has been deployed.
- We change the deploy path of the lobby to use product number instead of build
number. This means the lobby will be in a '2.6' folder instead of '2.6+build'
This means we will routinely overwrite the most recent lobby version until
we bump the product version. This is desirable so we have a limited number
of lobbies running.

* Remove version number from bot artifacts

* Change bot-version to be product version

Same product-versions will now overwrite previous bot installation
with the same product-version.

* Post-merge, update to use git-hash