Server: strip possible square brackets from 'getRemoteAddr' (#11038)
'getRemoteAddr' can potentially return IPv6 IP address with square
brackets around them. We need to strip these off for the IPv6 to
be a valid net address (otheriwse the DB will reject these values).

This update adds a utility that can be used to 'getRemoteAddr' and
have the square brackets be stripped.