Infra: always run 'ufw' commands (fix for --dry-run) (#11146)
When running with '--dry-run', "register" variables are not
defined. This causes the UFW tasks to fail when run in --dry-run
mode due to undefined variable error.

To resolve this, we remove the conditional run logic which removes
the usage of the register variable. The conditional logic was there
to avoid contention and sporadic failures when running the ufw command.
If we see failures again on UFW commands, we can re-evaluate then.