Configure nginx to send 2.5 clients to 2.6 (#11191)
This update would have NGINX (which is listening on 'prod.triplea-game.org')
send 2.5 game clients to the 2.6 lobby (2.5 is the current 'main' release).

The 2.6 lobby and 2.5 game clients are compatible with each other, so this
is a safe configuration update to make.

Though, 2.5 clients are currently configured to use a 'prod2' URL and they
would see no difference. But, we can configure the 2.5 clients to start
using the 'prod.triplea-game' server by modifying the lobby_games.yml file,
which would then cause 2.5 clients to be redirect to the 2.6 lobby. Until
then, this change has no impact.