expanded public static Image getImage(final Path path) errorMessage (#11231)
* expanded public static Image getImage(final Path path) errorMessage with:
"You should build with the checked in launcher, or and run 'downloadAssets'." to avoid unnecessary issues report because it's a dev issue. Typically, the logo image is bundled with the game, and if game bundling is a problem, this is not necessarily the first error you would see in that case (hence it is likely a dev issue).

* expanded public static Image getImage(final Path path) errorMessage with:
"You should build with the checked in launcher, or and run 'downloadAssets'." to avoid unnecessary issues report because it's a dev issue. Typically, the logo image is bundled with the game, and if game bundling is a problem, this is not necessarily the first error you would see in that case (hence it is likely a dev issue).