Test Speed: Faster smokes - check in XML file set (#11388)
Rather than downloading a lengthy list of XML files, this update commits a number of those XMLs directly into source control. This avoids the need to download the files in serial. Perhaps about 50% of the original list are now commited.

On my machine this reduced test execution speed of smoke tests from 40s to 33s.

+Small fix in TestDataFileListener.findFilesInFolder to account for the new location of the smoke-testing module in game-app/.