Build: update worfklow paths, build if 'build.gradle' is updated (#11496)
This update will have us build all of the workflows if the top-level
build.gradle file is updated. Notably we can have version changes
in the top level file. In our current configuration there is
no way to know which version changes are going to impact
which sub-module, so we need to build them all.