Fix NPE with returnFire field. (#11729)
The field was marked as transient in two classes, meaning it would not be serialized and was null on deserialization, but the code did not handle it properly. Since the leaf code using it was marked with  `@RemoveOnNextMajorRelease` and could handle null, simply remove the non-null tag and propagate the deprecation tag.

Fixes https://github.com/triplea-game/triplea/issues/11097.