Fix issue a cause of "Could not find step name" (2.6 bug). (#11773)
This was happening because the code to generate step names was not excluding units that would not participate in combat, resulting in infrastructure units getting their own steps (which later did not match what the engine generated once the filtering took place).

Uses the same logic as what's done for the battle to exclude units.

This change required adjusting a bunch of tests that were previously not careful about which had mistakes in setting up mock battles where the units didn't match the territory (in terms of sea vs. land). Also makes the tests to mock game data properties leniently, so that only the ones being set to true need to be specified (removing lots of LOC).

Includes a unit test.

Fixes: https://github.com/triplea-game/triplea/issues/10647
Note: Doesn't fix  https://github.com/triplea-game/triplea/issues/11617, as the root cause of that error is different, despite the actual error being the same. I will fix that in a separate PR.