Fix ArrayIndexOutOfBoundsException in CommentPanel. (#11786)
This a follow-up to my change that narrowed the scope of the lock in this code. We're now seeing node.getLastChild() throw an index out of bounds, even though it's directly after a size check. This is likely due to a race.

Instead of reading the children off the actual history tree, this changes the code to instead get this info directly from the event, which is more correct as well. It should fix the issue.