Fix another cause of "Could not find step name" (#11838)
The error would happen when there are multiple casualty steps at first, but when we get to that point in combat, some of those get removed (due units dying and not having a counterattack) and we may end up with a generic casualty step. To fix this, we keep track of which firing units each casualty step had and match based on those.

Adds a test for the fix that's confirmed to fail without it.

Note: This is fixing a 2.6 bug, so no release notes.