Fix "Could not find step name: withdraw?" error. (#11860)
This can happen when the withdraw step becomes available only when they are planes left. Since the UI doesn't get updated step names, it means switching to this step would produce the error.

Since switching to the step isn't really needed, since the action is via a dialog, omit the message to do so in this case.

Also, cleans up a bunch of related code.