Remove wiremock tests where we start a server during test (#12152)
This update removes the HTTP client tests for things like
the lobby, maps & game-support server. In these cases,
as part of testing, we launch these servers via docker-compose
and will then use our HTTP client to do testing. The independent
wiremock test of these same HTTP clients is redundant.

On the other hand, wiremock is still useful when we don't
control the server, or can't invoke the server all the time
(EG: githubs web-api). We still keep wiremock for this.