Fix a race on startup that could clear a loaded save. (#12437)
I noticed the race when running the game in debug mode from IntelliJ, where it was consistently being hit for me when I would click the button to open a save right after startup.
This PR fixes it by ensuring the save game will only be loaded after the initial map load.

Also, splits the function to load a map vs. a game and updates call sites.