Small clean up to sortUnitsForCasualtiesWithSupport(). (#12698)
Since we'll need to debug this code for https://github.com/triplea-game/triplea/issues/12689, this is a prelim cleanup to make the code easier to understand.

It splits out the core logic into an impl function, with the caching logic being a wrapper around it. Also improves some comments and uses some local variables to reduce wrapping.

No functional changes.