Automatic Bot Moderators (#12758)
Updates to give the host of a network game 'mod' powers. If the game is a bot, then also give the 'oldest' player 'mod' powers.

Mod powers allow:

    disconnect players
    ban players (effective until restart of bot)

The bot case is interesting since the 'oldest' player can change. For example, two players enter a bot (to the staging screen), first is mod & then leaves. Now the second player needs to become moderator.

To achieve this we add a "moderator promoted" message which is sent by the server to inform players of the new moderator.