Simpler Map unzip - delete old & simple unzip  (#12773)
Updates map extraction algorithm to the following:
  -  on startup, or immediately following map download: detect '.zip' file in 'downloadedMaps' folder
  -  determine target folder for extraction (strip '-master' suffix if present)
  -  delete target folder
  -  unzip the zip file to the target folder

The above is simpler notably because we just 'unzip'. Before the algorithm would check the case whether the extracted contents is a single folder, or if the zip contained all the map files in an exploded format. With this update, it does not matter, we just unzip into a destination folder.

The delete is unfortunate but required. Existing maps will potentially have this format: `[map-name]/*`, rather than something like `[map-name]/[map-name-mastter]/*`. Because of this mismatch, if we just extract, then we would wind up with duplicate map files.